RTGOV-480 Refactor to move actual provider implementations into shared war and hide mock implementations from platform specific wars. Fixed situation retrieval issue when switchyard jmx access returned 'instance not found' exception
