+++ modules/active-queries/active-collection-rests/.gitignore
@@ -1,0 +1,2 @@
+/target/
+/target/
