+++ ui/overlord-rtgov-ui-war-eap6/pom.xml
@@ -194,6 +194,7 @@
                 <exclude>WEB-INF/lib/jaxb-*.jar</exclude>
                 <exclude>WEB-INF/lib/netty*.jar</exclude>
                 <exclude>WEB-INF/lib/xercesImpl*.jar</exclude>
+                <exclude>WEB-INF/classes/org/overlord/rtgov/ui/server/services/mock/*</exclude>
               </excludes>
             </overlay>
           </overlays>
