+++ ui/overlord-rtgov-ui-war-fuse6/src/main/webapp/WEB-INF/beans.xml
@@ -2,9 +2,8 @@
 <beans xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xmlns:weld="http://jboss.org/schema/weld/beans" 
   xsi:schemaLocation="
-      http://java.sun.com/xml/ns/javaee
-      http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
-
+	  http://java.sun.com/xml/ns/javaee http://docs.jboss.org/cdi/beans_1_0.xsd
+      http://jboss.org/schema/weld/beans http://jboss.org/schema/weld/beans_1_1.xsd">
   <weld:scan>
     <weld:include name="org.overlord.**" />
   </weld:scan>
