+++ ui/overlord-rtgov-ui-war-jetty8/pom.xml
@@ -56,6 +56,7 @@
                 <exclude>WEB-INF/lib/rtgov-ui-situations-*.jar</exclude>
                 <exclude>WEB-INF/lib/situation-store-jpa-*.jar</exclude>
 
+                <exclude>WEB-INF/classes/org/overlord/rtgov/ui/server/services/mock/*</exclude>
               </excludes>
             </overlay>
           </overlays>
