+++ ui/overlord-rtgov-ui-war/src/main/webapp/WEB-INF/beans.xml
@@ -1,17 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://java.sun.com/xml/ns/javaee"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="
-      http://java.sun.com/xml/ns/javaee
-      http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
+       xmlns:weld="http://jboss.org/schema/weld/beans">
+
+	<weld:scan>
+		<weld:exclude name="org.overlord.rtgov.ui.server.services.impl.**" />
+	</weld:scan>
 
 	<interceptors>
 		<class>org.overlord.rtgov.ui.server.interceptors.UserContextInterceptor</class>
 	</interceptors>
 
     <alternatives>
-        <class>org.overlord.rtgov.ui.server.services.impl.MockSituationsServiceImpl</class>
-        <class>org.overlord.rtgov.ui.server.services.impl.MockServicesServiceImpl</class>
-        <class>org.overlord.rtgov.ui.server.services.impl.MockSituationEventGenerator</class>
+        <class>org.overlord.rtgov.ui.server.services.mock.MockSituationsServiceImpl</class>
+        <class>org.overlord.rtgov.ui.server.services.mock.MockServicesServiceImpl</class>
+        <class>org.overlord.rtgov.ui.server.services.mock.MockSituationEventGenerator</class>
     </alternatives>
 </beans>
